pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zero-trust-mcp-gateway"
version = "0.1.0"
description = "Zero-Trust Security Gateway for MCP and Agent Tool Calls"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

authors = [
  { name = "Maintainers", email = "dev@example.com" }
]

keywords = [
  "zero-trust",
  "mcp",
  "agent-security",
  "policy-enforcement",
  "llm-tools",
  "security-gateway",
  "model-context-protocol"
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Libraries",
  "Topic :: System :: Distributed Computing",
]

dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.1.0",
  "mypy>=1.0",
  "build>=0.10.0",
  "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/rajasblack/zero-trust-mcp-gateway"
Issues = "https://github.com/rajasblack/zero-trust-mcp-gateway/issues"
Documentation = "https://github.com/rajasblack/zero-trust-mcp-gateway#readme"

# -------------------------
# Packaging (src layout)
# -------------------------
[tool.setuptools]
packages = ["zero_trust_mcp_gateway"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
zero_trust_mcp_gateway = ["py.typed"]

# -------------------------
# Ruff (lint + format)
# -------------------------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["zero_trust_mcp_gateway"]

# -------------------------
# Mypy (optional strictness)
# -------------------------
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
